requirements.txt
# 새로운 기술 스택 - 핵심 패키지만
ultralytics>=8.0.0  # YOLOv8
openai>=1.0.0  # GPT-4 API
git+https://github.com/openai/CLIP.git  # CLIP model

# 기본 ML/AI 라이브러리
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# LangChain (GPT-4 체인용)
langchain>=0.1.0
langchain-openai>=0.1.0

# 컴퓨터 비전
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Streamlit 웹 프레임워크
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# 이미지 처리 및 시각화
matplotlib>=3.7.0
plotly>=5.15.0

# 유틸리티
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
pathlib2>=2.3.7

# 모델 훈련 추가 패키지
PyYAML>=6.0  # YOLO 설정 파일용
tensorboard>=2.13.0  # 훈련 모니터링
scikit-learn>=1.3.0  # 데이터 분할 및 평가

# CLIP for image content validation
openai-clip
ftfy
regex

# PDF Report Generation
reportlab>=4.0.0
weasyprint>=60.0
matplotlib>=3.7.0 